+++ packages/giu-examples/pages/datatable.js
@@ -619,7 +619,7 @@ class AppWrapper extends React.Component {
           />
           <link
             rel="stylesheet"
-            href={`${baseUrl}/static/font-awesome/css/font-awesome.min.css`}
+            href={`${baseUrl}/static/deps/font-awesome/css/font-awesome.min.css`}
           />
           <link
             rel="icon"