+++ packages/giu-examples/pages/material.js
@@ -387,19 +387,19 @@ class AppWrapper extends React.Component {
           />
           <link
             rel="stylesheet"
-            href={`${baseUrl}/static/font-awesome/css/font-awesome.min.css`}
+            href={`${baseUrl}/static/deps/font-awesome/css/font-awesome.min.css`}
           />
           <link
             rel="stylesheet"
-            href={`${baseUrl}/static/typeface-gloria-hallelujah/index.css`}
+            href={`${baseUrl}/static/deps/typeface-gloria-hallelujah/index.css`}
           />
           <link
             rel="stylesheet"
-            href={`${baseUrl}/static/typeface-roboto/index.css`}
+            href={`${baseUrl}/static/deps/typeface-roboto/index.css`}
           />
           <link
             rel="stylesheet"
-            href={`${baseUrl}/static/material-design-lite/material.min.css`}
+            href={`${baseUrl}/static/deps/material-design-lite/material.min.css`}
           />
           <link
             rel="stylesheet"