+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 * **Breaking**:
-    * Move 99.99% of styles to giu.css, so that they can be fully configured by the user.
+    * Move 99.99% of styles to **giu.css**, so that they can be fully configured by the user.
     * Remove `style` props.
     * Remove `zIndex` props.
+    * Remove **reset.css**.
 * Bump babel.
 * Remove tests + related dependencies.